code: fuzzy picker is scrollable!
- Fuzzy picker is scrollable. This works by storing the index of the
bottommost item to display in the fuzzy picker, and when the user tries
to select and item off the picker active page, we update that value to
"scroll".
- `gatewaySlack.requestConnectionUrl` returns an error now instead of
throwing one itself. I was running into a panic that happened because
this function still paniced instead of passing erros up the chain.